requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0
requests==2.31.0
pydantic==2.6.0
itsdangerous==2.1.2
